+++ src/lib.rs
@@ -14,27 +14,36 @@ extern crate thiserror;
 
 pub mod dachshund;
 
+pub use dachshund::adjacency_matrix::AdjacencyMatrix;
+pub use dachshund::algebraic_connectivity::AlgebraicConnectivity;
 pub use dachshund::beam::Beam;
+pub use dachshund::betweenness::Betweenness;
 pub use dachshund::candidate::Candidate;
+pub use dachshund::clustering::Clustering;
 pub use dachshund::cnm_communities::CNMCommunities;
 pub use dachshund::connected_components::ConnectedComponents;
+pub use dachshund::connectivity::Connectivity;
 pub use dachshund::coreness::Coreness;
+pub use dachshund::eigenvector_centrality::EigenvectorCentrality;
 pub use dachshund::graph_base::GraphBase;
 pub use dachshund::graph_builder::GraphBuilder;
 pub use dachshund::id_types::{EdgeTypeId, GraphId, NodeId, NodeTypeId};
 pub use dachshund::input::Input;
+pub use dachshund::laplacian::Laplacian;
 pub use dachshund::line_processor::LineProcessor;
 pub use dachshund::node::Node;
 pub use dachshund::output::Output;
 pub use dachshund::row::EdgeRow;
 pub use dachshund::scorer::Scorer;
 pub use dachshund::search_problem::SearchProblem;
+pub use dachshund::shortest_paths::ShortestPaths;
 pub use dachshund::simple_transformer::SimpleTransformer;
 pub use dachshund::simple_undirected_graph::SimpleUndirectedGraph;
 pub use dachshund::simple_undirected_graph_builder::SimpleUndirectedGraphBuilder;
 pub use dachshund::test_utils::*;
 pub use dachshund::transformer::Transformer;
 pub use dachshund::transformer_base::TransformerBase;
+pub use dachshund::transitivity::Transitivity;
 pub use dachshund::typed_graph::TypedGraph;
 pub use dachshund::typed_graph_builder::TypedGraphBuilder;
 pub use dachshund::typed_graph_line_processor::TypedGraphLineProcessor;
